+++ stats/pages_proofed_graphs.php
@@ -3,6 +3,7 @@ $relPath = './../pinc/';
 include_once($relPath.'base.inc');
 include_once($relPath.'theme.inc');
 include_once($relPath.'misc.inc'); // get_enumerated_param()
+include_once($relPath.'graph_data.inc');
 
 require_login();
 
@@ -10,19 +11,23 @@ $valid_tally_names = array_keys(get_page_tally_names());
 $tally_name = get_enumerated_param($_GET, 'tally_name', null, $valid_tally_names);
 
 $title = sprintf(_('Graphs for Pages Saved-as-Done in Round %s'), $tally_name);
-
-output_header($title);
-echo "<h1>$title</h1>";
-
-$images = [
-    "jpgraph_files/pages_daily.php?tally_name=$tally_name&amp;cori=increments&amp;timeframe=curr_month",
-    "jpgraph_files/pages_daily.php?tally_name=$tally_name&amp;cori=cumulative&amp;timeframe=curr_month",
-    "jpgraph_files/pages_daily.php?tally_name=$tally_name&amp;cori=increments&amp;timeframe=prev_month",
-    "jpgraph_files/pages_daily.php?tally_name=$tally_name&amp;cori=increments&amp;timeframe=all_time",
-    "jpgraph_files/pages_daily.php?tally_name=$tally_name&amp;cori=cumulative&amp;timeframe=all_time",
-    "jpgraph_files/total_pages_by_month_graph.php?tally_name=$tally_name",
+$graphs = [
+    ["barLineGraph", "pages_daily_increments_curr_month", pages_daily($tally_name, "increments", "curr_month")],
+    ["barLineGraph", "pages_daily_cumulative_curr_month", pages_daily($tally_name, "cumulative", "curr_month")],
+    ["barLineGraph", "pages_daily_increments_prev_month", pages_daily($tally_name, "increments", "prev_month")],
+    ["barLineGraph", "pages_daily_increments_all_time", query_graph_cache("pages_daily", [$tally_name, "increments", "all_time"])],
+    ["barLineGraph", "pages_daily_cumulative_all_time", pages_daily($tally_name, "cumulative", "all_time")],
+    ["barLineGraph", "total_pages_by_month_graph", total_pages_by_month_graph($tally_name)],
 ];
 
-foreach ($images as $image) {
-    echo "<img style='max-width: 100%' src='$image'></br>\n";
+output_header($title, SHOW_STATSBAR, [
+    "js_files" => get_graph_js_files(),
+    "js_data" => build_svg_graph_inits($graphs),
+]);
+echo "<h1>$title</h1>";
+
+echo "<div style='max-width: 640px'>";
+foreach ($graphs as [$type, $id]) {
+    echo "<div id='$id' style='max-height: 400px'></div><hr class='graph-divider'>";
 }
+echo "</div>";