+++ stats/projects_Xed_graphs.php
@@ -4,6 +4,7 @@ include_once($relPath.'base.inc');
 include_once($relPath.'project_states.inc');
 include_once($relPath.'theme.inc');
 include_once($relPath.'misc.inc'); // get_enumerated_param()
+include_once($relPath.'graph_data.inc');
 
 require_login();
 
@@ -11,17 +12,22 @@ $which = get_enumerated_param($_GET, 'which', null, $project_status_descriptors)
 
 $psd = get_project_status_descriptor($which);
 
-output_header($psd->graphs_title);
+$graphs = [
+    ["barLineGraph", "curr_month_proj", curr_month_proj($which)],
+    ["barLineGraph", "cumulative_month_proj", cumulative_month_proj($which)],
+    ["barLineGraph", "total_proj_graph", total_proj_graph($which)],
+    ["barLineGraph", "cumulative_total_proj_graph", cumulative_total_proj_graph($which)],
+];
+
+output_header($psd->graphs_title, SHOW_STATSBAR, [
+    "js_files" => get_graph_js_files(),
+    "js_data" => build_svg_graph_inits($graphs),
+]);
 
 echo "<h1>$psd->graphs_title</h1>";
 
-$images = [
-    "jpgraph_files/curr_month_proj.php?which=$which",
-    "jpgraph_files/cumulative_month_proj.php?which=$which",
-    "jpgraph_files/total_proj_graph.php?which=$which",
-    "jpgraph_files/cumulative_total_proj_graph.php?which=$which",
-];
-
-foreach ($images as $image) {
-    echo "<img style='max-width: 100%' src='$image'><br>\n";
+echo "<div style='max-width: 640px'>";
+foreach ($graphs as [$type, $id]) {
+    echo "<div id='$id' style='max-height: 400px'></div><hr class='graph-divider'>";
 }
+echo "</div>";