+++ stats/stats_central.php
@@ -6,11 +6,21 @@ include_once($relPath.'theme.inc');
 include_once($relPath.'ThemedTable.inc');
 include_once($relPath.'site_news.inc');
 include_once($relPath.'misc.inc');
+include_once($relPath.'graph_data.inc');
 
 require_login();
 
 $title = _("Statistics Central");
-output_header($title);
+
+$cumulative_total_proj_summary_graph_data = cumulative_total_proj_summary_graph();
+$graphs = [
+    ["stackedAreaGraph", "cumulative_total_proj_summary_graph", $cumulative_total_proj_summary_graph_data],
+];
+
+output_header($title, SHOW_STATSBAR, [
+    "js_files" => get_graph_js_files(),
+    "js_data" => build_svg_graph_inits($graphs),
+]);
 
 echo "<h1>" . _("Statistics Central") . "</h1>\n";
 
@@ -235,6 +245,21 @@ foreach (['created', 'proofed', 'PPd', 'posted'] as $which) {
 $table->end();
 
 // XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX
+// Miscellaneous Graphs
+
+$table = new ThemedTable(
+  2,
+  _("Miscellaneous Graphs")
+);
+
+$table->row(
+  "<a href='equilibria.php'>" . _("Pages Saved per Day (Equilibria)") . "</a>",
+  "<a href='pages_in_states.php'>" . _("Pages in States") . "</a>"
+);
+
+$table->end();
+
+// XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX
 // Total Projects Created, Proofread, Post-Processed and Posted
 
 $table = new ThemedTable(
@@ -246,11 +271,8 @@ $table->row(
     '&nbsp;'
 );
 
-$img_url = "jpgraph_files/cumulative_total_proj_summary_graph.php";
-$alt_text = _("Total Projects Created, Proofread, Post-Processed and Posted");
-
 $table->row(
-    "<img style='max-width: 100%' src='$img_url' alt='" . attr_safe($alt_text) . "'>"
+    "<div id='cumulative_total_proj_summary_graph' style='max-width: 640px'></div>"
 );
 
 $table->end();